+++ src/blocks/core.clj
@@ -8,8 +8,8 @@
   When blocks are returned from a block store, they may include 'stat' metadata
   about the blocks, including:
 
+  - `:source`      resource location for the block content
   - `:stored-at`   time block was added to the store
-  - `:origin`      resource location for the block
   "
   (:refer-clojure :exclude [get list])
   (:require
@@ -49,6 +49,7 @@
      (data/lazy-block id (.length file) reader))))
 
 
+; TODO: support opening a byte range
 (defn open
   "Opens an input stream to read the content of the block. Throws an IO
   exception on empty blocks."
@@ -69,7 +70,7 @@
   ([source]
    (read! source default-algorithm))
   ([source algorithm]
-   (data/read-block source algorithm)))
+   (data/read-block algorithm source)))
 
 
 (defn write!