+++ CHANGELOG.md
@@ -9,11 +9,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ...
 
+## [0.4.1] - 2015-11-12
+
+### Changed
+- Rename `:origin` block stat to `:source`.
+- Switch argument order in `read-block` for consistency.
+
+### Fixed
+- `put!` retains extra attributes and metadata on the block argument in the
+  returned block.
+- Expanded integration test suite to cover `stat` and `get` on non-existent
+  blocks and `put!` merging.
+
 ## [0.4.0] - 2015-11-10
 
-Lots of high-level (breaking) library changes! `blocks.data.Block` is now a
-custom type to protect immutable fields like `:id` and `:size` and support the
-`IPending` interface.
+Lots of high-level library changes! `blocks.data.Block` is now a custom type to
+protect immutable fields like `:id` and `:size` and support the `IPending`
+interface.
 
 ### Added
 - Blocks can be either _literal_ or _lazy_ to support larger block sizes.
@@ -31,4 +43,6 @@ custom type to protect immutable fields like `:id` and `:size` and support the
 
 Initial project release.
 
-[Unreleased]: https://github.com/greglook/blocks/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/greglook/blocks/compare/0.4.1...HEAD
+[0.4.1]: https://github.com/greglook/blocks/compare/0.4.0...0.4.1
+[0.4.0]: https://github.com/greglook/blocks/compare/0.3.0...0.4.0